Replace wp/data with hooks using useApiFetch
