--- client/data/onboarding/action-types.ts
@@ -1,8 +1,0 @@
-/** @format */
-
-enum ACTION_TYPES {
-	SET_BUSINESS_TYPES = 'SET_BUSINESS_TYPES',
-	SET_REQUIRED_VERIFICATION_INFO = 'SET_REQUIRED_VERIFICATION_INFO',
-}
-
-export default ACTION_TYPES;
