--- client/data/onboarding/index.ts
@@ -1,12 +1,0 @@
-/** @format */
-
-/**
- * Internal dependencies
- */
-import reducer from './reducer';
-import * as selectors from './selectors';
-import * as actions from './actions';
-import * as resolvers from './resolvers';
-
-export { reducer, selectors, actions, resolvers };
-export * from './hooks';
