+++ client/data/store.js
@@ -17,8 +17,6 @@ import * as settings from './settings';
 import * as multiCurrency from './multi-currency';
 import * as readers from './card-readers';
 import * as capital from './capital';
-import * as documents from './documents';
-import * as onboarding from './onboarding';
 
 // Extracted into wrapper function to facilitate testing.
 export const initStore = () =>
@@ -33,8 +31,6 @@ export const initStore = () =>
 			multiCurrency: multiCurrency.reducer,
 			readers: readers.reducer,
 			capital: capital.reducer,
-			documents: documents.reducer,
-			onboarding: onboarding.reducer,
 		} ),
 		actions: {
 			...deposits.actions,
@@ -46,8 +42,6 @@ export const initStore = () =>
 			...multiCurrency.actions,
 			...readers.actions,
 			...capital.actions,
-			...documents.actions,
-			...onboarding.actions,
 		},
 		controls,
 		selectors: {
@@ -60,8 +54,6 @@ export const initStore = () =>
 			...multiCurrency.selectors,
 			...readers.selectors,
 			...capital.selectors,
-			...documents.selectors,
-			...onboarding.selectors,
 		},
 		resolvers: {
 			...deposits.resolvers,
@@ -73,7 +65,5 @@ export const initStore = () =>
 			...multiCurrency.resolvers,
 			...readers.resolvers,
 			...capital.resolvers,
-			...documents.resolvers,
-			...onboarding.resolvers,
 		},
 	} );
